+++ src/main/java/be/ugent/rml/records/ReferenceFormulation.java
@@ -1,0 +1,10 @@
+package be.ugent.rml.records;
+
+import be.ugent.rml.NAMESPACES;
+
+public class ReferenceFormulation {
+    public static String XPath = NAMESPACES.QL + "XPath";
+    public static String CSV = NAMESPACES.QL + "CSV";
+    public static String RDB = NAMESPACES.QL + "RDB";
+    public static String JSONPath = NAMESPACES.QL + "JSONPath";
+}