+++ src/test/java/be/ugent/rml/Mapper_MySQL_Test.java
@@ -98,7 +98,7 @@ public class Mapper_MySQL_Test extends MySQLTestCore {
                 {"RMLTC0012c", Error.class, BOTH},
                 {"RMLTC0012d", Error.class, BOTH},
                 {"RMLTC0012e", null, BOTH},
-//                {"RMLTC0013a", null, BOTH},
+                {"RMLTC0013a", null, BOTH},
                 {"RMLTC0014d", null, BOTH},
 //                {"RMLTC0015a", null, BOTH},
                 {"RMLTC0015b", Error.class, BOTH},